Move GenerationRunner to giselle-engine package
