+++ apps/playground/package.json
@@ -17,7 +17,6 @@
 	"dependencies": {
 		"@giselle-internal/workflow-designer-ui": "workspace:*",
 		"@giselle-sdk/data-type": "workspace:*",
-		"@giselle-sdk/generation-runner": "workspace:^",
 		"@giselle-sdk/giselle-engine": "workspace:*",
 		"@giselle-sdk/run": "workspace:^",
 		"@giselle-sdk/workflow-designer": "workspace:*",
