--- packages/generation-runner/package.json
@@ -1,44 +1,0 @@
-{
-	"name": "@giselle-sdk/generation-runner",
-	"version": "0.0.0",
-	"private": true,
-	"main": "./dist/index.js",
-	"module": "./dist/index.mjs",
-	"types": "./dist/index.d.ts",
-	"scripts": {
-		"build": "tsup",
-		"check-types": "tsc --noEmit",
-		"clean": "rm -rf dist && rm -rf test/dist",
-		"format": "biome check --write ."
-	},
-	"exports": {
-		"./package.json": "./package.json",
-		".": {
-			"types": "./dist/index.d.ts",
-			"import": "./dist/index.mjs",
-			"require": "./dist/index.js",
-			"default": "./dist/index.js"
-		},
-		"./react": {
-			"types": "./react/dist/index.d.ts",
-			"import": "./react/dist/index.mjs",
-			"require": "./react/dist/index.js",
-			"default": "./react/dist/index.js"
-		},
-		"./react-internal": "./src/react/index.ts"
-	},
-	"devDependencies": {
-		"@giselle/giselle-sdk-tsconfig": "workspace:*",
-		"@types/react": "catalog:",
-		"tsup": "catalog:"
-	},
-	"dependencies": {
-		"@ai-sdk/react": "catalog:",
-		"@giselle-sdk/data-type": "workspace:^",
-		"@giselle-sdk/giselle-engine": "workspace:^",
-		"@giselle-sdk/telemetry": "workspace:^",
-		"ai": "catalog:",
-		"react": "catalog:",
-		"zod": "catalog:"
-	}
-}
