--- packages/generation-runner/src/generation-runner.ts
@@ -1,4 +1,0 @@
-export function GenerationRunner() {
-	function createGeneration() {}
-	function startGeneration() {}
-}
