--- packages/generation-runner/tsup.config.ts
@@ -1,17 +1,0 @@
-import { defineConfig } from "tsup";
-
-export default defineConfig([
-	{
-		entry: ["src/index.ts"],
-		format: ["cjs", "esm"],
-		dts: true,
-		sourcemap: true,
-	},
-	{
-		entry: ["src/react/index.ts"],
-		format: ["cjs", "esm"],
-		dts: true,
-		outDir: "react/dist",
-		sourcemap: true,
-	},
-]);
