+++ packages/giselle-engine/package.json
@@ -55,10 +55,12 @@
 		"@ai-sdk/google": "catalog:",
 		"@ai-sdk/openai": "catalog:",
 		"@ai-sdk/perplexity": "catalog:",
+		"@ai-sdk/react": "catalog:",
 		"@fal-ai/client": "catalog:",
 		"@giselle-sdk/data-mod": "workspace:^",
 		"@giselle-sdk/github-client": "workspace:^",
 		"@giselle-sdk/language-model": "workspace:^",
+		"@giselle-sdk/telemetry": "workspace:^",
 		"@giselle-sdk/text-editor": "workspace:^",
 		"@giselle-sdk/usage-limits": "workspace:^",
 		"@giselle-sdk/workflow-utils": "workspace:^",
