+++ packages/giselle-engine/src/react/generations/contexts/generation-runner-system.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import type { Message } from "@ai-sdk/react";
 import {
 	type CancelledGeneration,
@@ -16,7 +18,6 @@ import {
 	isFailedGeneration,
 	isRunningGeneration,
 } from "@giselle-sdk/data-type";
-import { useGiselleEngine } from "@giselle-sdk/giselle-engine/react";
 import {
 	type ReactNode,
 	createContext,
@@ -26,13 +27,14 @@ import {
 	useRef,
 	useState,
 } from "react";
+import { useGiselleEngine } from "../../use-giselle-engine";
+import { GenerationRunner } from "../generation-runner";
 import {
 	arrayEquals,
 	waitAndGetGenerationCompleted,
 	waitAndGetGenerationFailed,
 	waitAndGetGenerationRunning,
-} from "../../helpers";
-import { GenerationRunner } from "../generation-runner";
+} from "../helpers";
 
 interface StartGenerationOptions {
 	onGenerationCreated?: (generation: CreatedGeneration) => void;
