+++ packages/workflow-designer/package.json
@@ -37,7 +37,6 @@
 	},
 	"dependencies": {
 		"@giselle-sdk/data-type": "workspace:*",
-		"@giselle-sdk/generation-runner": "workspace:^",
 		"@giselle-sdk/giselle-engine": "workspace:*",
 		"@giselle-sdk/language-model": "workspace:^",
 		"@giselle-sdk/run": "workspace:^",
