+++ packages/workspace/package.json
@@ -33,7 +33,6 @@
 	},
 	"dependencies": {
 		"@giselle-sdk/data-type": "workspace:^",
-		"@giselle-sdk/generation-runner": "workspace:^",
 		"@giselle-sdk/giselle-engine": "workspace:^",
 		"@giselle-sdk/integration": "workspace:^",
 		"@giselle-sdk/run": "workspace:^",
