+++ packages/workspace/src/react/hooks.ts
@@ -1,4 +1,4 @@
-export { useNodeGenerations } from "@giselle-sdk/generation-runner/react";
+export { useNodeGenerations } from "@giselle-sdk/giselle-engine/react";
 export { useRun, useRunController } from "@giselle-sdk/run/react";
 export { useUsageLimits } from "@giselle-sdk/usage-limits/react";
 export { useWorkflowDesigner } from "@giselle-sdk/workflow-designer/react";
