+++ packages/workspace/src/react/workspace.tsx
@@ -1,8 +1,10 @@
 "use client";
 
 import type { Workspace, WorkspaceId } from "@giselle-sdk/data-type";
-import { GenerationRunnerSystemProvider } from "@giselle-sdk/generation-runner/react";
-import { useGiselleEngine } from "@giselle-sdk/giselle-engine/react";
+import {
+	GenerationRunnerSystemProvider,
+	useGiselleEngine,
+} from "@giselle-sdk/giselle-engine/react";
 import type { Integration } from "@giselle-sdk/integration";
 import { IntegrationProvider } from "@giselle-sdk/integration/react";
 import { RunSystemContextProvider } from "@giselle-sdk/run/react";
